remove support for fd_driven
Actually supporting it would require to modify the activity to
pass the fd-set as exceptionfds instead of read-fds. Since
I don't have a use of this right now (the GPIOs I interface with
don't generate interrupts), I'm passing on finishing the
functionality.

One can always revert this commit once support can be implemented.