fix: fix heisenbug and simplify tests
A heisenbug was happening in the "does not keep last written value when a new connection is established with 'init: false'" test. The test expected no samples, but got 1.
This was due to how the test was structured, specifically the fact that the write and the disconnect-connect block of code were inside the same expect_execution block, causing a race condition.

The tests were also simplified by using a data_reader instead of the previous, unnecessary complicated, setup.